[IMP] test_mail(_full): cleanup mail push tests
Web push tests are moved into 'test_mail'as it is now part of standard mail
code. 'test_mail_full' is used for integrations with advanced mail modules
like sms or mass mailing.

A quick cleaning is done on push tests
  * split into sub-tests to clarify input / output and better spot use cases;
  * clean code;
  * clean setup data;
  * perform some renaming;

Task-3669738
