+++ addons/test_mail_full/tests/__init__.py
@@ -8,4 +8,3 @@ from . import test_mass_mailing
 from . import test_portal
 from . import test_rating
 from . import test_res_users
-from . import test_web_push
